# docs/swagger.yaml
basePath: /
definitions:
  models.CreatePetBody:
    properties:
      petName:
        type: string
      specieID:
        type: string
    required:
    - petName
    - specieID
    type: object
  models.RegisterAnUserBody:
    properties:
      email:
        example: ronald@email.com
        type: string
      password:
        example: aksdmalknj@154/JKNJ
        type: string
      username:
        example: ronald123
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    description: User represents the model  for an User's Data, password is ALWAYS
      HASHED
    properties:
      email:
        example: ronald@email.com
        type: string
      id:
        description: |-
          Users ID generated by Mongodb
          In: path
        example: "1"
        type: string
      money:
        example: 500000
        type: integer
      password:
        example: aksdmalknj@154/JKNJ
        type: string
      pets:
        items:
          type: string
        type: array
      username:
        example: ronald123
        type: string
      vivarium:
        items:
          type: string
        type: array
    required:
    - email
    - password
    - username
    type: object
  models.UserCredentials:
    properties:
      password:
        example: aksdmalknj@154/JKNJ
        type: string
      username:
        example: ronald123
        type: string
    required:
    - password
    - username
    type: object
  responses.Response:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: lucasl22l@proton.me
    name: API Support
  description: This is the API for setting the REST functions of the Virtual Pets,
    JWT in Swag has no good suppert . so sadly the AUTH is not working and need a
    migration
  termsOfService: http://swagger.io/terms/
  title: Virtual Pets
  version: "1.0"
paths:
  /api/pets:
    post:
      consumes:
      - application/json
      description: Creates a pet with a specific Species and a Name passed by the
        user
      operationId: create-a-pet
      parameters:
      - description: Create a Pet
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.CreatePetBody'
      produces:
      - application/json
      responses:
        "200":
          description: Pet linked successfuly
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: This pet has other owner!
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: User/Pet not found
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure on the Server
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - jwt: []
      summary: Create a Pet
      tags:
      - Pet
  /api/users/{username}:
    get:
      description: Receives Details from a specific User (without password)
      operationId: get-an-user-profile
      parameters:
      - description: User's name
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Details
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure on the Server
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get an User Details (Profile)
      tags:
      - User
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Login an User with a Username and Password
      operationId: login-an-user
      parameters:
      - description: Logs an User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: Logged User and return JWT cookie
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure in the Database
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Login an User
      tags:
      - User
  /api/users/logout:
    post:
      description: Logout current logged in user by expiring his Jwt auth
      operationId: logout-an-user
      produces:
      - application/json
      responses:
        "200":
          description: Logout successfuly!
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure on the Server
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Logout an User
      tags:
      - User
  /api/users/pet/{petID}:
    put:
      description: When creating a pet successfully, this functions link to the user
        that it created based on the JWT
      operationId: link-a-pet-to-user
      parameters:
      - description: Pet ID
        in: path
        name: petID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pet linked successfuly
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: This pet has other owner!
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: User/Pet not found
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure on the Server
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - jwt: []
      summary: Link a created Pet to an User
      tags:
      - User
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register a new User with a Username, Email and Password
      operationId: register-a-user
      parameters:
      - description: Register a new User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterAnUserBody'
      produces:
      - application/json
      responses:
        "201":
          description: User Created!
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Invalid Body
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Failure in the Database
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Register a new User
      tags:
      - User
swagger: "2.0"